Worked in Google Docs
Made IDs for all the options and the items. Sorted the IDs be making the
option IDs 5 numbers long and in the font syncopate. The item IDs are 3
numbers long and in the font Cutive Mono. The story is in the font Comic
Sans MS. Went through and made all the changes to font as necessary.